player run and enemy AI

+++ Assets/Scripts/DamageSystem/Damage.cs
@@ -4,13 +4,13 @@ namespace DamageSystem
 {
     public class Damage : IDamage
     {
-        public int Value { get; }
+        public float Value { get; }
         public DamageType Type { get; }
         public LayerMask Layer { get; }
         
         public void CollideWith(IDamageable damageable){}
 
-        public Damage(int value, DamageType type, LayerMask layer = default)
+        public Damage(float value, DamageType type, LayerMask layer = default)
         {
             Value = value;
             Type = type;
