+++ Assets/Scripts/DamageSystem/Hitbox.cs
@@ -5,11 +5,11 @@ namespace DamageSystem
 {
     public class Hitbox: MonoBehaviour, IDamage
     {
-        [SerializeField] private int _damageValue;
+        [SerializeField] private float _damageValue;
         [SerializeField] private DamageType _damageType;
         [SerializeField] private LayerMask _layerMask;
         
-        public int Value => _damageValue;
+        public float Value => _damageValue;
         public DamageType Type => _damageType;
         public LayerMask Layer => _layerMask;
