+++ Assets/Scripts/DamageSystem/IDamage.cs
@@ -4,7 +4,7 @@ namespace DamageSystem
 {
     public interface IDamage
     {
-        public int Value { get; }
+        public float Value { get; }
         public DamageType Type { get; }
         public LayerMask Layer { get; }
